Fix passing None to combine
In a custom holiday rule, if the end date when calling businesstimedelta.difference is greater than the last holiday in the rule then a TypeError is returned because None is returned by HolidayRule.next. This can be fixed by returning a date in the future that is not plausible. i.e. 9999-12-31